modified to work with output from updated

+++ graphfun/data/graph_grappler.py
@@ -24,8 +24,9 @@ def get_graphs(df: pd.DataFrame,
     """
     graphs = df['graph6_rep'].to_list()
     
-    labels = np.array(df['planar'].to_list(), dtype=int)
-
+    #labels = np.array(df['planar'].to_list(), dtype=int)
+    labels = np.array(df['planar'].to_list())
+    #G = [nx.from_graph6_bytes(g.strip().encode()) for g in graphs]
     G = [nx.from_graph6_bytes(g.strip().encode()) for g in graphs]
 
     if return_adj: return np.array(graphList_to_adj(G)), labels
